+++ examples/show_random_chart.py
@@ -4,17 +4,17 @@ import random
 import sys
 
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
-from tchart.tchart import ChartRenderer
+from tchart import Tchart
 
 
 def main():
-    r = ChartRenderer(height=10, width=120)
+    t = Tchart(height=10, width=120)
     data = []
 
-    for i in range(random.randint(10, 100)):
+    for _ in range(random.randint(10, 100)):
         data.append(random.uniform(-100, 100))
 
-    chart = r.render(data)
+    chart = t.render(data)
     print('\n'.join(chart))